Re #17839 Overhaul system test for pearl powder
